Modified scanner wrapper
Modified scanner wrapper to pass a 'meta-job' ID to ORTClient. This will
allow us to identify when multiple jobs that appear server-side should
be grouped together. The ID is unique to the current can process; it
gets passed to all child scan jobs and does not get passed to any jobs
once a new scan process has started.