Added User Metadata ID
Bringing back 'job-metadata-id' as 'user-metadata-id'. The root ORT
instance will create a unique metadata id that can be used by the first
call to ORT2LiAn to read the contents of an additional metadata file and
push those to Mongo.

Subsequent calls to ORT2Lian with a user-metadata-id supplied but no
additional metadata file to query Mongo to get the additional data. The
call can block for a given length of time to ensure the first ORT2LiAn
instance records the data before the first read.